refactor: 리뷰 피드백 반영
