+++ src/main/java/roomescape/controller/ReservationController.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
-import roomescape.entity.Reservation;
+import roomescape.domain.Reservation;
 
 @Controller
 public class ReservationController {
